Implement lal upgrade and a daily, silent upgrade check.
This is very unobtrusive, and it will just say what to do if there's a new version
found in the artifactory storage with a greater semver version. lal upgrade will work
provided you are using the musl install, otherwise it will tell you to pull and cargo build.

Last updateTime is stored in the lalrc and is updated on the daily update check.