Add dynamic completion of lal script
This is a breaking change to current `.lal/scripts` folders as it `lal` now sources and executes the `main` function in such a script directly rather than calling the script. Thus the scripts are assumed to be bash libraries.

This allows the functions to export a function called `completer` which can return a one-level completion for this script.
Typically this information would live together with the script so this makes sense.

This is a bit magic, but it totally works, and it's really really nice. Corresponding changes to relevant script repos upcoming.