implement verify environment safeguards
Pretty neat little refactor in lockfile (apart from the non-dynamic
get_value) to allow the recursive function we already wrote to fetch
environments. May not work for all types of keys, but string keys like
currently, yeah, this is fine.

this closes #23
This also closes #24 as build now calls out to verify correctly as well
(and shell + script already done)