fix tests and build in configured build container
this is a bit hacky in the tests, but have documented it
we are entering the rust container, but are verifying centos deps
artificially. not really that bad.

because the rust container is obtained from the ~/.lal/config, we can
now provision this properly for the build-workers and not hardcode it
anywhere. Thus this fixes #39