handle env based on command type + implement multi fetch
This closes #22 + #26 + #33 + #27

There's also quite a bit of refactoring work in main() to try and tidy
up what env means for different commands. I.e. #42

Currently, it's now split into ones that:

- need the proper env values: (dockery ones + verify + env subcommands)
- need explicit defaults when nothing is set (everything in install:
  fetch, update, update-all as well as export and query)
- are env agnostic (status/ls, listers, rm, stash)

The special cases are still handled in main, but the rest are factored
into functions. Main was getting cyclotomically large anyway.