provide better error message for first io operation in build
This was a common case error for people on nfs mounts and the error was
not very good. At least it will tell you that your normal user cannot
touch OUTPUT now. There are other file operations later, but they are
mostly in the same directory, and there's associated trace messages.

Thus, either lal build gives you an obvious error, or `lal -vv build`
should allow you to pinpoint it better. Unfortunately, it's not 100%
perfect. There's no deep source of which exact IO operation failed in
the io::Error (and there really should be), so this is the best we can
do for now.

This closes #44