lal publish basic working version + artifactory auth support
This is the first working version of #45 but it needs to get the
checksum headers in there.

Restructured the artifactory struct once more to indicate more clearly
what we use what for...

Was a bit of a learning curve doing this with hyper, but got the hang of it now. :)