add a timestamp to lockfile (temporarily optional)
this closes #52 because adding it to the version key for lal stash
looked ugly. we may present the timestamp in lal ls -t later anyway.

lal stash now edits the lockfile in OUTPUT so that the value is
persisted in the lockfile in ~/.lal/cache/stash rather than just in the
folder name.