allow manifest to be in .lal subfolder
this becomes the new default and there's a enum to keep track of how we
choose. in a codebase with multiple manifests, we use the subfoldered
one, but warn if there are duplicate (should only be seen in those cases
or if there's user errors)

This is for #68

lal init --public is scrapped, always write to default for now.