lal fetch now doesn't re-extract already present deps
this speeds up `lal fetch` to basically be a noop in most cases.
we basically cross check what we find in the full lockfile, and take out
(from the list of things to install) things that are already there, at
the right version, and built in the right environment.

we no longer aggressively `rm -rf INPUT` at the start of fetch as a
consequence, but we have started doing it if `lal fetch` fails (because
the INPUT folder could look valid and have lockfiles, but have partial
extracts).

This closes #73