factor out edonus default environments from the rust code
`lal configure` now takes a default file, so people can use it with
other environments. for now this is only available in source in the
configs/ directory, but should probably be supplied in the built
artifactory tarball as well and have some nice autocompletions for the
available environments.

This is a initial sketch of #59 - keeping it in a branch because it may
need some bikeshedding.