add files to tarball without leading ./ paths
this was broken by #78 and was __technically__ breaking, but only
jenkins and it's pedantic ways would notice. regardless this is fixing
it and will be released in the next patch version.

thanks to @ppettina for how to `man tar` better.

ended up also upgrading walkdir while i was trying things out, works, so
doing it.