make lockfile.environment required
this was always set by lal>=2, but this breaks compat with components
built before that. thus doing the initial bump to 3.0.0-alpha here.

this also now expects that manifest.environment exists so #83 should be
done for better error handling, and components must be gone through before we can merge this.

additionally this revealed an inconsistency in environment agnostic
commands so #85 should be looked at, as well as #81, and #84 for
clarity.