write a Backend trait for artifactory - closes #84
This writes down the assumptions we expect from artifactory in a trait
(which is kind of like an abstract mixin).

It also makes environment a `std::Option` where it is allowed so no
default strings more than necessary (still a few in artifactory, but
will clear those up).

This is also for #47