implement CachedBackend for all Backends
Have basically moved all the weird artifactory assumptions in 3.0 into one artifactory.rs file and codified what we need in a `Backend` trait (in traits.rs), so that any replacement would only need to implement this trait. `CachedBackend` is auto-implemented for any implementors of the `Backend` trait, and most commands can just use this instead (only a few commands need `Backend` directly:  [publish, upgrade, query] - so pretty happy with this abstraction).

Configuration for artifactory is also wrapped behind a
`BackendConfiguration` enum (with only one struct entry).

This means even the config doesn't really know about Artifactory, and
everything can be switched over at some point if necessary. Default
impluementation of the BackendConfiguration was implemented so that we
only needed one line in main to decide it.