make autoupgrade optional - for #61
because it can be considered invasive, allow users who know what they
are doing to turn it off. it's just curl taring the lastest url anyway.
could be done in a cronjob or even you bashrc if you want to write a
small conditional for it.

it's also useful to jenkins because it avoids the future-date hack to do
the same thing (set lastUpgrade way in the future kind of does the same
thing).

Changes the config structure slightly again, but this is for the
breaking 3.0.0