make autoupgrade a compile feature
probably want to disable this for potential general use.
it also did make some logic in main.rs a bit less dense (particularly
we factored something out of the main function).

Now `upgrade` subcommand does not exist (runtime check) if the feature
was not compile in, and the functions related to its handling are not
compiled in when the feature is off.