prevent lal from running as root - for #132
Currently only check on configure as a start. People can short circuit this by having a config ready, but in the future we can re-use this error case in build for a more rigorous check. It's generally unnecessary to sudo docker build anyway, but it doesn't work without remapping setup.