enforce component names are normal lowercase
originally started this to avoid invalid paths, but turns out we can
have a path of unicode characters like; `ὀδυσσεύς`.

non-unicode paths in cache should be impossible because the manifest
only allows strings (unicode) in the manifest anyway.

however, this is checking that the component names are lowercase,
because don't want to have case sensitivity there.

thus a component name like: `ὈΔΥΣΣΕΎΣ` is disallowed.

the lowercase check is thus put everywhere. maybe we want to upgrade it
to ascii lowercase only though..